Modify Dockerfile to run sshd
This branch is intended for demonstration purposes only. Please
don't merge it into the master/main branch.

This is an example of what a Dockerfile for the frontend console
container would look like:

* it would need to run sshd
* it would need authorised persons public SSH keys

We would need to 'docker build' both this image and the existing
image used for serving traffic / e2e testing.