Add configuration file, model and tests
What

Add a YAML configuration file that will be used to inform us when
a specific service (via it's LGSL code) is not available in a
particular country.

Add a model that reads the config file and provides methods to easily
access the various parts of the configuration.

Add tests for the above model.

Why

Currently the user interface does not inform the user that a service
is unavailable in a particular country. They are usually taken to
the county website with a message saying that this is what we have
matched their postcode to. This is not a good user experience.

We want to inform the user when a service is unavailable for a specific
country or countries.