Add new page for unavailable services
What

Add a new page with suitable messaging, URL etc when a service is unavailable in a particular country.

Highlight the postcode the user has entered.

Why

So that the user is better informed that the service they are looking for
is unavailable at the postcode given and provide them with an appropriate alternative resource that might be able to assist them with finding a suitable alternate or country specific service.